+++ api/v2/dogu.go
@@ -9,13 +9,18 @@ import (
 // You cannot add dogus which are not yet in the blueprint.
 type MaskDogu struct {
 	// Name defines the name of the dogu including its namespace, f. i. "official/nginx". Must not be empty.
+	// +required
 	Name string `json:"name"`
 	// Version defines the version of the dogu that is to be installed. Must not be empty if the targetState is "present";
 	// otherwise it is optional and is not going to be interpreted.
-	Version string `json:"version"`
+	// +optional
+	Version *string `json:"version,omitempty"`
 	// Absent defines if the dogu should be absent in the ecosystem. Defaults to false.
-	Absent bool `json:"absent,omitempty"`
+	// +optional
+	Absent *bool `json:"absent,omitempty"`
 }
+
+// +kubebuilder:validation:XValidation:message="dogu version must not be empty",rule="(!has(self.absent) || !self.absent) ? has(self.version) : true"
 
 // Dogu defines a Dogu, its version, and the installation state in which it is supposed to be after a blueprint
 // was applied.
@@ -25,12 +30,14 @@ type Dogu struct {
 	Name string `json:"name"`
 	// Version defines the version of the dogu that is to be installed. Must not be empty if the targetState is "present";
 	// otherwise it is optional and is not going to be interpreted.
-	// +required
-	Version string `json:"version"`
+	// +optional
+	Version *string `json:"version,omitempty"`
 	// Absent defines if the dogu should be absent in the ecosystem. Defaults to false.
-	Absent bool `json:"absent,omitempty"`
+	// +optional
+	Absent *bool `json:"absent,omitempty"`
 	// PlatformConfig defines infrastructure configuration around the dogu, such as reverse proxy config, volume size etc.
-	PlatformConfig PlatformConfig `json:"platformConfig,omitempty"`
+	// +optional
+	PlatformConfig *PlatformConfig `json:"platformConfig,omitempty"`
 }
 
 func (in *Dogu) DeepCopy() *Dogu {
@@ -53,13 +60,17 @@ func (in *Dogu) DeepCopyInto(out *Dogu) {
 }
 
 type ResourceConfig struct {
-	MinVolumeSize string `json:"minVolumeSize,omitempty"`
+	// +optional
+	MinVolumeSize *string `json:"minVolumeSize,omitempty"`
 }
 
 type ReverseProxyConfig struct {
-	MaxBodySize      string `json:"maxBodySize,omitempty"`
-	RewriteTarget    string `json:"rewriteTarget,omitempty"`
-	AdditionalConfig string `json:"additionalConfig,omitempty"`
+	// +optional
+	MaxBodySize *string `json:"maxBodySize,omitempty"`
+	// +optional
+	RewriteTarget *string `json:"rewriteTarget,omitempty"`
+	// +optional
+	AdditionalConfig *string `json:"additionalConfig,omitempty"`
 }
 
 type DataSourceType string
@@ -78,18 +89,24 @@ type AdditionalMount struct {
 	// Valid options are:
 	//   ConfigMap - data stored in a kubernetes ConfigMap.
 	//   Secret - data stored in a kubernetes Secret.
+	// +required
 	SourceType DataSourceType `json:"sourceType"`
 	// Name is the name of the data source.
+	// +required
 	Name string `json:"name"`
 	// Volume is the name of the volume to which the data should be mounted. It is defined in the respective dogu.json.
+	// +required
 	Volume string `json:"volume"`
 	// Subfolder defines a subfolder in which the data should be put within the volume.
 	// +optional
-	Subfolder string `json:"subfolder,omitempty"`
+	Subfolder *string `json:"subfolder,omitempty"`
 }
 
 type PlatformConfig struct {
-	ResourceConfig         ResourceConfig     `json:"resource,omitempty"`
-	ReverseProxyConfig     ReverseProxyConfig `json:"reverseProxy,omitempty"`
-	AdditionalMountsConfig []AdditionalMount  `json:"additionalMounts,omitempty" patchStrategy:"replace"`
+	// +optional
+	ResourceConfig *ResourceConfig `json:"resource,omitempty"`
+	// +optional
+	ReverseProxyConfig *ReverseProxyConfig `json:"reverseProxy,omitempty"`
+	// +optional
+	AdditionalMountsConfig []AdditionalMount `json:"additionalMounts,omitempty" patchStrategy:"replace"`
 }