+++ api/v2/dogu_test.go
@@ -1,11 +1,14 @@
 package v2
 
 import (
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestDogu_DeepCopy(t *testing.T) {
+	falsePtr := false
+	version1234 := "1.2.3-4"
 	tests := []struct {
 		name  string
 		given *Dogu
@@ -15,13 +18,13 @@ func TestDogu_DeepCopy(t *testing.T) {
 			name: "empty",
 			given: &Dogu{
 				Name:    "test",
-				Version: "1.2.3-4",
-				Absent:  false,
+				Version: &version1234,
+				Absent:  &falsePtr,
 			},
 			want: &Dogu{
 				Name:    "test",
-				Version: "1.2.3-4",
-				Absent:  false,
+				Version: &version1234,
+				Absent:  &falsePtr,
 			},
 		},
 	}