+++ api/v2/globalConfigDiff.go
@@ -1,15 +1,7 @@
 package v2
 
 // GlobalConfigDiff is a list of differences between the GlobalConfig in the Blueprint and the cluster state
-type GlobalConfigDiff []GlobalConfigEntryDiff
+type GlobalConfigDiff []ConfigEntryDiff
 
 // GlobalConfigValueState represents either the actual or expected state of a global config key
 type GlobalConfigValueState ConfigValueState
-
-// GlobalConfigEntryDiff contains the difference and the needed actions for a single global config key
-type GlobalConfigEntryDiff struct {
-	Key          string                 `json:"key"`
-	Actual       GlobalConfigValueState `json:"actual"`
-	Expected     GlobalConfigValueState `json:"expected"`
-	NeededAction ConfigAction           `json:"neededAction"`
-}